+++ CHANGELOG.md
@@ -15,6 +15,9 @@ See [0Ver](https://0ver.org/).
   decorator is applied with explicit exception types
 - Add picky exceptions to `impure_safe` decorator like `safe` has. Issue #1543
 - Add partition function to result module. Issue #1905
+- Adds `default_error` parameter to `returns.converters.maybe_to_result`,
+  which provides a default error value for `Failure`
+
 
 ### Misc