+++ CHANGELOG.md
@@ -6,16 +6,19 @@ incremental in minor, bugfixes only are patches.
 See [0Ver](https://0ver.org/).
 
 
-## 0.24.0 WIP
+## 0.24.0
 
 ### Features
 
+- Drop `python3.9` support
+- Add `python3.13` support
+- Add support for `mypy>=1.12`
 - Add picky exceptions to `future_safe` decorator like `safe` has.
 - Improve inference of `ResultLike` objects when exception catching
   decorator is applied with explicit exception types
 - Add picky exceptions to `impure_safe` decorator like `safe` has. Issue #1543
 - Add partition function to result module. Issue #1905
-- Adds `default_error` parameter to `returns.converters.maybe_to_result`,
+- Add `default_error` parameter to `returns.converters.maybe_to_result`,
   which provides a default error value for `Failure`